Update ssl.pcap
The current pcap is a bit rough -- it's cooked-mode encapsulated, which means
that tcpdump by default doesn't render it too well ...

  $ tcpdump -s0 -r ssl.pcap  | head
  reading from file ssl.pcap, link-type LINUX_SLL (Linux cooked v1), snapshot length 262144
  [Invalid header: len(2656) < caplen(2672)]
  [Invalid header: len(2656) < caplen(2672)]
  [Invalid header: len(464) < caplen(480)]

... and the contents aren't particularly interesting. This replacement has three
HTTPS connections, each requesting google.com, once with TLS 1.2 and a valid
cert as of today, once with TLS 1.3, and once with TLS 1.2 far in the future so
Zeek considers the cert expired. Checksums are correct.